+++ internal/pkg/util/bin/bin_test.go
@@ -3,6 +3,8 @@
 // LICENSE.md file distributed with the sources of this project regarding your
 // rights to use or distribute this software.
 
+//go:build singularity_engine
+
 package bin
 
 import (
